requirements.txt
Django>=4.2
djangorestframework
pymongo
python-dotenv
PyJWT
bcrypt
pandas
geopandas
python-dateutil
gunicorn
